--- scripts/helpers/database.js
@@ -1,9 +1,0 @@
-module.exports = function saveToCollection(collection, item, res) {
-  db.collection(collection).save(item, (err, result) => {
-    if (err) {
-      return console.log(err);
-    };
-
-    res.redirect('/urls');
-  });
-}
