+++ scripts/helpers/generateRandomId.js
@@ -1,0 +1,4 @@
+module.exports = function generateRandomId() {
+  let randomId = Math.random().toString(36).substring(5);
+  return randomId;
+}
