+++ views/partials/_delete.ejs
@@ -1,3 +1,3 @@
 <form action="/urls/<%= shortUrl %>/delete" method="POST" class="<%= className %>">
-  <input type="submit" value="Delete">
+  <button type="submit">Delete</button>
 </form>
